renamed run.py to train.py, tested historical self-play
