--- run.py
@@ -1,241 +1,0 @@
-from pdb import set_trace as T
-import importlib
-import argparse
-import inspect
-import yaml
-import time
-import sys
-
-import pufferlib
-import pufferlib.utils
-
-from reinforcement_learning import clean_pufferl, environment
-
-DEBUG = False
-# See curriculum_generation/manual_curriculum.py for details
-BASELINE_CURRICULUM = "curriculum_generation/curriculum_with_embedding.pkl"
-
-
-def load_from_config(agent, debug=False):
-    with open('config.yaml') as f:
-        config = yaml.safe_load(f)
-    default_keys = 'agent_zoo env train policy recurrent sweep_metadata sweep_metric sweep wandb postproc'.split()
-    defaults = {key: config.get(key, {}) for key in default_keys}
-
-    debug_config = config.get('debug', {}) if debug else {}
-    agent_config = config[agent]
-
-    combined_config = {}
-    for key in default_keys:
-        agent_subconfig = agent_config.get(key, {})
-        debug_subconfig = debug_config.get(key, {})
-        combined_config[key] = {**defaults[key], **agent_subconfig, **debug_subconfig}
-
-    return pufferlib.namespace(**combined_config)
-
-def get_init_args(fn):
-    if fn is None:
-        return {}
-    sig = inspect.signature(fn)
-    args = {}
-    for name, param in sig.parameters.items():
-        if name in ('self', 'env', 'policy'):
-            continue
-        if name in ('agent_id', 'is_multiagent'):  # Postprocessor args
-            continue
-        if param.kind == inspect.Parameter.VAR_POSITIONAL:
-            continue
-        elif param.kind == inspect.Parameter.VAR_KEYWORD:
-            continue
-        else:
-            args[name] = param.default if param.default is not inspect.Parameter.empty else None
-    return args
-
-def init_wandb(args, resume=True):
-    if args.no_track:
-        return None
-    assert args.wandb.project is not None, 'Please set the wandb project in config.yaml'
-    assert args.wandb.entity is not None, 'Ple ase set the wandb entity in config.yaml'
-    import wandb
-    wandb_kwargs = {
-        'id': args.exp_name or wandb.util.generate_id(),
-        'project': args.wandb.project,
-        'entity': args.wandb.entity,
-        'config': {
-            'cleanrl': args.train,
-            'env': args.env,
-            'agent_zoo': args.agent,
-            'policy': args.policy,
-            'postproc': args.postproc,
-            'recurrent': args.recurrent,
-        },
-        'name': args.exp_name,
-        'monitor_gym': True,
-        'save_code': True,
-        'resume': resume,
-    }
-    if args.wandb.group is not None:
-        wandb_kwargs['group'] = args.wandb.group
-    return wandb.init(**wandb_kwargs)
-
-# Return env_creator, agent_creator
-def setup_agent(module_name):
-    try:
-        agent_module = importlib.import_module(f'agent_zoo.{module_name}')
-    except ModuleNotFoundError:
-        raise ValueError(f'Agent module {module_name} not found under the agent_zoo directory.')
-    env_creator = environment.make_env_creator(postprocessor_cls=agent_module.Postprocessor)
-    def agent_creator(env, args):
-        policy = agent_module.Policy(env, **args.policy)
-        # if args.force_recurrence or env_module.Recurrent is not None:
-        #     policy = env_module.Recurrent(env, policy, **args.recurrent)
-        #     policy = pufferlib.frameworks.cleanrl.RecurrentPolicy(policy)
-        # else:
-        policy = pufferlib.frameworks.cleanrl.Policy(policy)
-        return policy.to(args.train.device)
-    init_args = {
-        'policy': get_init_args(agent_module.Policy.__init__),
-        'postproc': get_init_args(agent_module.Postprocessor.__init__),
-    }
-    return agent_module, env_creator, agent_creator, init_args
-
-def train(args, env_creator, agent_creator):
-    data = clean_pufferl.create(
-        config=args.train,
-        agent_creator=agent_creator,
-        agent_kwargs={'args': args},
-        env_creator=env_creator,
-        env_creator_kwargs={'env': args.env, 'postproc': args.postproc},
-        vectorization=args.vectorization,
-        exp_name=args.exp_name,
-        track=args.track,
-    )
-
-    while not clean_pufferl.done_training(data):
-        clean_pufferl.evaluate(data)
-        clean_pufferl.train(data)
-
-    print('Done training. Saving data...')
-    clean_pufferl.close(data)
-    print('Run complete.')
-
-def sweep(args, env_creator, agent_creator):
-    import wandb
-    sweep_id = wandb.sweep(sweep=args.sweep, project=args.wandb.project)
-
-    def main():
-        try:
-            args.exp_name = init_wandb(args).id
-            if hasattr(wandb.config, 'train'):
-                # TODO: Add update method to namespace
-                print(args.train.__dict__)
-                print(wandb.config.train)
-                args.train.__dict__.update(dict(wandb.config.train))
-            train(args, env_creator, agent_creator)
-        except Exception as e:
-            import traceback
-            traceback.print_exc()
-
-    wandb.agent(sweep_id, main, count=20)
-
-if __name__ == '__main__':
-    parser = argparse.ArgumentParser(description='Parse environment argument', add_help=False)
-    parser.add_argument('-m', '--mode', type=str, default='train', choices='train sweep evaluate'.split())
-    parser.add_argument('-a', '--agent', type=str, default='neurips23_start_kit', help='Agent module to use')
-    parser.add_argument('-n', '--exp-name', type=str, default=None, help="Need exp name to resume the experiment")
-    parser.add_argument('-p', '--eval-model-path', type=str, default=None, help='Path to model to evaluate')
-    parser.add_argument('-c', '--curriculum', type=str, default=BASELINE_CURRICULUM, help='Path to curriculum file')
-    #parser.add_argument('--baseline', action='store_true', help='Baseline run')
-    #parser.add_argument('--no-render', action='store_true', help='Disable render during evaluate')
-    parser.add_argument('--vectorization', type=str, default='multiprocessing', choices='serial multiprocessing ray'.split())
-    parser.add_argument('--force-recurrence', action='store_true', help='Force model to be recurrent, regardless of defaults')
-    if DEBUG:
-        parser.add_argument('--no-track', default=True, help='Do NOT track on WandB')
-        parser.add_argument('--debug', default=True, help='Debug mode')
-    else:
-        parser.add_argument('--no-track', action='store_true', help='Do NOT track on WandB')
-        parser.add_argument('--debug', action='store_true', help='Debug mode')
-
-    clean_parser = argparse.ArgumentParser(parents=[parser])
-    args = parser.parse_known_args()[0].__dict__
-    config = load_from_config(args['agent'], debug=args.get('debug', False))
-    agent_module, env_creator, agent_creator, init_args = setup_agent(args['agent'])
-
-    # Update config with environment defaults
-    config.policy = {**init_args['policy'], **config.policy}
-    config.postproc = {**init_args['postproc'], **config.postproc}
-    config.recurrent = {}  # {**get_init_args(env_module.Recurrent.__init__), **config.recurrent}
-
-    # Generate argparse menu from config
-    for name, sub_config in config.items():
-        args[name] = {}
-        for key, value in sub_config.items():
-            data_key = f'{name}.{key}'
-            cli_key = f'--{data_key}'.replace('_', '-')
-            if isinstance(value, bool) and value is False:
-                action = 'store_false'
-                parser.add_argument(cli_key, default=value, action='store_true')
-                clean_parser.add_argument(cli_key, default=value, action='store_true')
-            elif isinstance(value, bool) and value is True:
-                data_key = f'{name}.no_{key}'
-                cli_key = f'--{data_key}'.replace('_', '-')
-                parser.add_argument(cli_key, default=value, action='store_false')
-                clean_parser.add_argument(cli_key, default=value, action='store_false')
-            else:
-                parser.add_argument(cli_key, default=value, type=type(value))
-                clean_parser.add_argument(cli_key, default=value, metavar='', type=type(value))
-
-            args[name][key] = getattr(parser.parse_known_args()[0], data_key)
-        args[name] = pufferlib.namespace(**args[name])
-
-    clean_parser.parse_args(sys.argv[1:])
-    args = pufferlib.namespace(**args)
-
-    # Args update: curriculum, vectorization, track, exp_name
-    args.env.curriculum_file_path = args.curriculum
-    vec = args.vectorization
-    if vec == 'serial' or args.debug:
-        args.vectorization = pufferlib.vectorization.Serial
-    elif vec == 'multiprocessing':
-        args.vectorization = pufferlib.vectorization.Multiprocessing
-    elif vec == 'ray':
-        args.vectorization = pufferlib.vectorization.Ray
-    else:
-        raise ValueError('Invalid --vectorization (serial/multiprocessing/ray).')
-
-    args.track = not args.no_track
-    if args.track:
-        args.exp_name = init_wandb(args).id
-    else:
-        args.exp_name = f"nmmo_{time.strftime('%Y%m%d_%H%M%S')}"
-    # elif args.baseline:
-    #     args.track = True
-    #     version = '.'.join(pufferlib.__version__.split('.')[:2])
-    #     args.exp_name = f'puf-{version}-nmmo'
-    #     args.wandb_group = f'puf-{version}-baseline'
-    #     shutil.rmtree(f'experiments/{args.exp_name}', ignore_errors=True)
-    #     run = init_wandb(args, resume=False)
-    #     if args.mode == 'evaluate':
-    #         model_name = f'puf-{version}-nmmo_model:latest'
-    #         artifact = run.use_artifact(model_name)
-    #         data_dir = artifact.download()
-    #         model_file = max(os.listdir(data_dir))
-    #         args.eval_model_path = os.path.join(data_dir, model_file)
-
-    if args.mode == 'train':
-        train(args, env_creator, agent_creator)
-        exit(0)
-    elif args.mode == 'sweep':
-        sweep(args, env_creator, agent_creator)
-        exit(0)
-    elif args.mode == 'evaluate':
-        clean_pufferl.rollout(
-            env_creator,
-            env_creator_kwargs={'env': args.env, 'postproc': args.postproc},
-            agent_creator=agent_creator,
-            agent_kwargs={'args': args},
-            model_path=args.eval_model_path,
-            device=args.train.device
-        )
-    else:
-        raise ValueError('Mode must be one of train, sweep, or evaluate')
